refactor: entities 구조 변경
폴더 내부에 파일이 많지 않으므로 폴더 뎁스 줄이기 위해 하나의 폴더경로로 통일
-> 결론적으로 최대 2개의 뎁스가 감소함
