+++ src/entities/cartItem/response.ts
@@ -1,7 +1,7 @@
 import {
   BaseProduct,
   PaginatedResponse,
-} from "../../../../shared/api/types/response";
+} from "../../shared/api/types/response";
 
 export interface CartItemContent {
   id: number;
