+++ src/entities/product/response.ts
@@ -1,6 +1,6 @@
 import {
   BaseProduct,
   PaginatedResponse,
-} from "../../../../shared/api/types/response";
+} from "../../shared/api/types/response";
 
 export type Products = PaginatedResponse<BaseProduct>;
