+++ src/widgets/ProductCatalog/ProductControlPanel/ProductSorter/ProductSorter.tsx
@@ -1,4 +1,4 @@
-import { useProducts } from "../../../../entities/product/model/hooks/useProducts";
+import { useProducts } from "../../../../entities/product/useProducts";
 import { sortOptions } from "../../../../shared/config/filter";
 import SelectBox from "../../../../shared/ui/SelectBox/SelectBox";
